must fix some tests
